@@ -19,7 +19,7 @@
         "submit"
     ],
     "engines": {
-        "vscode": "^1.30.0"
+        "vscode": "^1.44.0"
     },
     "author": {
         "name": "Marcelo Fornet"
@@ -200,9 +200,10 @@
     "scripts": {
         "vscode:prepublish": "npm run compile",
         "compile": "tsc -p ./",
+        "lint": "prettier --check src",
         "watch": "tsc -watch -p ./",
-        "postinstall": "node ./node_modules/vscode/bin/install",
-        "test": "npm run compile && node ./node_modules/vscode/bin/test"
+        "pretest": "npm run compile",
+        "test": "node ./out/test/runTest.js"
     },
     "repository": {
         "type": "git",
@@ -213,23 +214,32 @@
         "email": "mfornet94@gmail.com"
     },
     "devDependencies": {
-        "@types/got": "^9.3.0",
-        "@types/mocha": "^2.2.42",
-        "@types/node": "^8.10.25",
+        "@types/glob": "^7.1.1",
         "@types/md5-file": "^4.0.0",
+        "@types/mocha": "^7.0.2",
+        "@types/node": "^13.11.0",
+        "@types/vscode": "^1.44.0",
+        "@typescript-eslint/eslint-plugin": "^2.26.0",
+        "@typescript-eslint/parser": "^2.26.0",
+        "@types/tmp": "^0.1.0",
+        "@types/rimraf": "^3.0.0",
+        "@types/express": "^4.16.4",
+        "@types/body-parser": "^1.19.0",
+        "@types/fs-extra": "^8.1.0",
+        "tmp": "^0.1.0",
+        "eslint": "^6.8.0",
+        "glob": "^7.1.6",
+        "mocha": "^7.1.1",
         "tslint": "^5.8.0",
-        "typescript": "^3.1.4",
-        "vscode": "^1.1.33"
+        "typescript": "^3.8.3",
+        "vscode-test": "^1.3.0"
     },
     "dependencies": {
-        "got": "^9.5.0",
-        "jssoup": "0.0.10",
-        "sync-request": "^6.0.0",
-        "unescape": "^1.0.1",
         "express": "^4.16.4",
-        "body-parser": "^1.18.3",
+        "body-parser": "^1.19.0",
         "md5-file": "^4.0.0",
         "clipboardy": "^2.2.0",
-        "rimraf": "^3.0.2"
+        "rimraf": "^3.0.2",
+        "fs-extra": "^9.0.0"
     }
 }
